[Xamarin]Refactor YogaKit
